Se arreglaron problemas en la asignacion de nodos azules que causeaba que a veces se asignaran nodos duplicados o que no correspondieran a nodos del grafo

+++ orange.h
@@ -150,7 +150,7 @@ class Orange{
         void processJoinRequest(Orange*, PacketEntry*);
         void respondToBlueRequest(Orange*, Token*);
         void initBlueMap();
-        int findNextUnassigned(Orange*);
+        unsigned short int findNextUnassigned(Orange*);
         
 
     public:
